requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
ta>=0.11.0
arch>=6.0.0
scipy>=1.11.0
structlog>=23.1.0
aiosqlite>=0.19.0
websockets>=12.0
python-telegram-bot>=20.0
click>=8.1.0
tabulate>=0.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
